% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{add.confounder.variables}
\alias{add.confounder.variables}
\title{Add confounder variable to environment object}
\usage{
add.confounder.variables(environment, ...)
}
\arguments{
\item{environment}{The environment object}

\item{...}{confounding variables}
}
\description{
TODO
}
