% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess2.R
\name{add_cell_score}
\alias{add_cell_score}
\title{Add cell specific score to metadata}
\usage{
add_cell_score(sce, name, values)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{name}{The name of the variable}

\item{values}{The value of the variables}
}
\description{
Add cell specific score to metadata
}
