% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simiarlity.R
\name{assess.cluster.similarity}
\alias{assess.cluster.similarity}
\title{Assess Cluster Similarity}
\usage{
assess.cluster.similarity(environment,
  diff.exp.file = "main.datasets.diff.exp.rds",
  cluster.similarity.function = pearson.correlation, label = "pearson",
  rerun = F)
}
\arguments{
\item{environment}{\code{environment} object}

\item{diff.exp.file}{name of differential expression results file}

\item{cluster.similarity.function}{which similarity function to use (either 'pearson.correlation' or '?') Mamie - there was another similarity function using euclidean distance. Do you know where did it go to? Can you replace the '?' with the name of this other function?}

\item{label}{name of the similarity measure to use for the results folder}

\item{rerun}{whether to rerun the analysis or load from cache}
}
\value{
pairwise cluster similarity measures
}
\description{
Assess similarity between pairs of clusters.
}
