% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{cell.cycle.score}
\alias{cell.cycle.score}
\title{Compute cell cycle score}
\usage{
cell.cycle.score(environment, knn = 10, cc.genes.path = NA)
}
\arguments{
\item{environment}{The environment object}

\item{knn}{The number of nearest neighbor}

\item{cc.genes.path}{Optional; path to defined cell cycle genes. Default used TODO (citation)}
}
\value{
TODO
}
\description{
TODO
}
