% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess2.R
\name{cell_cycle_score}
\alias{cell_cycle_score}
\title{Compute Cell Cycle Score}
\usage{
cell_cycle_score(sce, knn = 10, cc_cycle_genes = NULL, verbose = F)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{knn}{number of nearest neighbor}

\item{verbose}{Whether to print diagnostic messages}

\item{cc_genes_genes}{cell cycle gene set. Default uses gene sets defined in Kowalczyk, M. S. et al. Single-cell RNA-seq reveals changes in cell cycle and differentiation programs upon aging of hematopoietic stem cells. Genome Res 25, 1860-1872, doi:10.1101/gr.192237.115 (2015).}
}
\value{
a matrix of cell cycle genes activation scores (S, G2M and aggregated S/G2M scores, separately)
}
\description{
Compute the activation of cell cycle genes defined in Kowalczyk, M. S. et al.
}
