% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{cluster.analysis}
\alias{cluster.analysis}
\title{Clustering analysis}
\usage{
cluster.analysis(environment, knn.ratios = c(0.01, 0.05, 0.1),
  nShuffleRuns = 10, shuffledKNN = 10, loadPreviousKnn = T,
  rerun = F, deleteCache = F, mem = "4GB", time = "0:15:00",
  plot = T)
}
\arguments{
\item{knn.ratios}{Range of KNN parameters to scan (corresponding to different resolutions)}

\item{nShuffleRuns}{Number of shuffled clustering analyses to perform per KNN threshold}

\item{loadPreviousKnn}{Whether to load previous analysis results}

\item{rerun}{Whether to rerun}

\item{mem}{HPC memory}

\item{time}{HPC time}

\item{plot}{Whether to plot the clustering qualities compared to shuffled}

\item{shuffleKNN}{Number of closest KNN shuffled analyses to include in background clustering quality computation}

\item{deleteCacheWhether}{to delete cache files}
}
\value{
TODO
}
\description{
TODO
}
