% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess2.R
\name{controlled_mean_score}
\alias{controlled_mean_score}
\title{Compute Controlled Activation Score}
\usage{
controlled_mean_score(rbs, genes, knn = 10, exclude.missing.genes = T,
  constrain.cell.universe = NULL, verbose = F)
}
\arguments{
\item{rbs}{A RobustSingleCell object}

\item{genes}{gene list upon which to calculate gene signature activate}

\item{knn}{number of nearest neighbors}

\item{exclude.missing.genes}{whether to exclude genes with missing values}

\item{constrain.cell.universe}{binary vector indicating in which subset of cells to calculate the gene signature activation. Default is all cells.}
}
\value{
gene signature activation scores per cell
}
\description{
Compute mean gene signatures activation scores while controlling for technically similar genes.
}
