% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_LCMV}
\alias{download_LCMV}
\title{Download Example Dataset}
\usage{
download_LCMV(base_dir)
}
\arguments{
\item{base_dir}{Full path to a directory where data and analysis will be stored}
}
\description{
Download two replicates of CD44+ T cell 10X scRNAseq data sets (Ciucci 2018).
}
\examples{
\dontrun{
download_LCMV(tempdir()) # require network connection
}
}
