% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{filter.cluster.data}
\alias{filter.cluster.data}
\title{Remove selected clusters}
\usage{
\method{filter}{cluster.data}(environment, remove.clusters)
}
\arguments{
\item{environment}{The \code{environment} object}

\item{remove.clusters}{A character vector of the clusters to be removed}
}
\value{
An environment object with selected clusters removed
}
\description{
Remove selected clusters from the environment object.
}
