% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data2.R
\name{filter}
\alias{filter}
\title{Filter expression matrix}
\usage{
filter(sce, min.genes.per.cell = 500,
  max.genes.per.cell.quantile = 0.98,
  max.UMIs.per.cell.quantile = 0.98, min.cells.per.gene = 1,
  max.mitochondrial.frac = 0.1, max.ribosomal.frac = NULL,
  verbose = F)
}
\arguments{
\item{min.genes.per.cell}{minimum required number of genes per cell}

\item{max.genes.per.cell.quantile}{upper quantile for number of genes per cell}

\item{max.UMIs.per.cell.quantile}{upper quantile for number of UMIs per cell}

\item{min.cells.per.gene}{minimum required number of cells per gene}

\item{max.mitochondrial.frac}{maximum fraction of reads mapped to mitochondrial
genes per cell}

\item{max.ribosomal.frac}{maximum fraction of reads mapped to ribosomal genes per cell}

\item{environment}{SingleCellExperiment object}

\item{genome}{genome annotation}
}
\value{
A filtered SingleCellExperiment object
}
\description{
Filter expression matrix
}
