% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{filter_cluster_data}
\alias{filter_cluster_data}
\title{Remove selected clusters}
\usage{
filter_cluster_data(environment, remove.clusters)
}
\arguments{
\item{environment}{The \code{environment} object}

\item{remove.clusters}{A character vector of the clusters to be removed}
}
\value{
An environment object with selected clusters removed
}
\description{
Remove selected clusters from the environment object.
}
\examples{
LCMV1 <- setup_LCMV_example()
LCMV1 <- filter_cluster_data(LCMV1, "1")
}
