% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get.cluster.names}
\alias{get.cluster.names}
\title{Get Cluster Names by Marker Gene Expression}
\usage{
get.cluster.names(environment, types, min.fold = 1.25, max.Qval = 0.1,
  print = T)
}
\arguments{
\item{environment}{\code{environment} object}

\item{types}{data frame associating cell type or state with marker genes}

\item{min.fold}{minimum fold change to consider a marker as overexpressed}

\item{max.Qval}{maximum FDR q value to consider a marker as overexpressed}

\item{print}{whether to print output calculations}
}
\value{
vector containing assigned cluster name labels
}
\description{
Use predefined marker genes to assign clusters with putative cell type or state labels.
}
