% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simiarlity.R
\name{get.robust.cluster.similarity}
\alias{get.robust.cluster.similarity}
\title{Get Robust Cluster Similarity}
\usage{
get.robust.cluster.similarity(environment, similarity,
  min.sd = qnorm(0.95), max.q.val = 0.01, rerun = F)
}
\arguments{
\item{environment}{The environment object}

\item{similarity}{TODO}

\item{min.sd}{TODO}

\item{max.q.val}{TODO}

\item{rerun}{TODO}
}
\description{
TODO
}
