% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess2.R
\name{get_variable_genes}
\alias{get_variable_genes}
\title{Identify Highly Variable Genes}
\usage{
get_variable_genes(sce, min.mean = 0.05, min.frac.cells = 0,
  min.dispersion.scaled = 1, plot = F, verbose = F)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{min.mean}{minimum mean expression per gene}

\item{min.frac.cells}{minimum fraction of cells expressing each gene}

\item{min.dispersion.scaled}{minimum dispersion value}

\item{plot}{whether to plot mean vs dispersion plot}

\item{verbose}{whether to print diagnostic message}
}
\value{
\code{SingleCellExperiment} parameter containing highly variable genes selection
}
\description{
Get highly variable genes by Heteroscedasticity controlled binning of gene expression measurements within each dataset separately.
}
