% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize.project}
\alias{initialize.project}
\title{Intialize the project object}
\usage{
initialize.project(datasets, origins, experiments, data.path, work.path,
  marker.genes = NULL, clear.history = F, analysis.label = NA,
  convert.to.mouse.gene.symbols = NULL)
}
\arguments{
\item{datasets}{List of dataset code names}

\item{origins}{List of dataset tissue origin/condition full name}

\item{experiments}{List of experiment design annotations}

\item{data.path}{Path to where the data is located}

\item{work.path}{path to where the analysis should create folders and store files}

\item{marker.genes}{Set of genes of interest for visualization purposes}

\item{clear.history}{Whether you would like to remove any previous project by this name}

\item{analysis.label}{Whether you would like to add a specific label to the analysis folder}

\item{convert.to.mouse.gene.symbols}{Whether you are using human gene symbols and would like to convert them
to mouse gene symbols}
}
\value{
\code{environment} parameter containing file paths and experiment parameters
}
\description{
Intialize the project object
}
