% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LCMV1_normalized}
\alias{LCMV1_normalized}
\title{The normalized expression matrix of LCMV1 dataset}
\format{A matrix with 1408 genes (rows) and 97 cells.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM3423795}
}
\usage{
LCMV1_normalized
}
\description{
A processed scRNAseq dataset containing randomly subsampld cells and highly
variable genes.
}
\keyword{datasets}
