% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LCMV1_small}
\alias{LCMV1_small}
\title{An processed environment object from 10X scRNAseq dataset}
\format{An object containing a matrix with 1408 genes (rows) and 97 cells.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM3423795}
}
\usage{
LCMV1_small
}
\description{
A processed scRNAseq dataset containing randomly subsampld cells and highly
variable genes of CD44+ T cells.
}
\keyword{datasets}
