% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LCMV2_normalized}
\alias{LCMV2_normalized}
\title{The normalized expression matrix of LCMV2 dataset}
\format{A matrix with 989 genes (rows) and 96 cells.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM3423794}
}
\usage{
LCMV2_normalized
}
\description{
A processed scRNAseq dataset containing randomly subsampld cells and highly
variable genes.
}
\keyword{datasets}
