% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data2.R
\name{log_normalize}
\alias{log_normalize}
\title{Log normalize the expression matrix}
\usage{
log_normalize(sce, verbose = F)
}
\arguments{
\item{sce}{A SingleCellExperiment object}
}
\value{
Return a SingleCellExperiment object
}
\description{
Log normalize the expression matrix
}
