% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{mitochondrial.score}
\alias{mitochondrial.score}
\title{Compute mitochondrial score}
\usage{
mitochondrial.score(environment, control = F, knn = 10)
}
\arguments{
\item{environment}{The environment object}

\item{control}{TODO}

\item{knn}{Number of nearest neighbor}
}
\value{
A vector of mitochondrial score (TODO)
}
\description{
TODO
}
