% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess2.R
\name{mitochondrial_score}
\alias{mitochondrial_score}
\title{Compute Mitochondrial Score}
\usage{
mitochondrial_score(sce, control = F, knn = 10, force = F,
  verbose = F)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{control}{whether to subtract the score defined by technically similar genes}

\item{knn}{number of nearest neighbor}

\item{force}{Whether to rerun and overwrite already computed mitochondrial score}

\item{verbose}{Whether to print diagnostic messages}
}
\value{
a vector of mitochondrial genes activation score
}
\description{
Compute the activation level of mitochondrial genes.
}
\examples{
LCMV1 <- setup_LCMV_example()
mitochondrial.score <- mitochondrial.score(LCMV1)
}
