% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.PCA}
\alias{plot.PCA}
\title{Plot PCA results as cell embedding in 2D space and annotation of gene loadings}
\usage{
\method{plot}{PCA}(environment, quantile, order)
}
\arguments{
\item{environment}{The environment object}

\item{quantile}{quantile}

\item{order}{order}
}
\description{
Plot PCA results as cell embedding in 2D space and annotation of gene loadings
}
