% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pair_scatter}
\alias{plot_pair_scatter}
\title{Plot Pairwise Gene Scatter Plot}
\usage{
plot_pair_scatter(environment, gene1, gene2, cluster_group1,
  cluster_group2, group1_label, group2_label, width = 10, height = 10)
}
\arguments{
\item{environment}{\code{environment} object}

\item{gene1}{selected gene number 1}

\item{gene2}{selected gene number 2}

\item{cluster_group1}{cluster group 1 to be visualized (one or more clusters)}

\item{cluster_group2}{cluster group 2 to be visualized (one or more clusters)}

\item{group1_label}{label for group 1 legend and file name}

\item{group2_label}{label for group 2 legend and file name}

\item{width}{pdf file canvas width}

\item{height}{pdf file canvas height}
}
\description{
Visualize normalized expression contours of a selected gene pair across selected cluster groups.
}
\examples{
\donttest{
plot_pair_scatter (environment,
gene1 = 'Cd4',
gene2 = 'Cd8',
cluster_group1 = c('cluster_name_1','cluster_name_2'),
cluster_group2 = c('cluster_name_3','cluster_name_4'),
group1_label = 'CD4 T Cells',
group2_label = 'CD8 T Cells')
}
}
