% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Plot PCA results}
\usage{
plot_PCA(environment, quantile, order)
}
\arguments{
\item{environment}{\code{environment} object}

\item{quantile}{quantile of PCA loadings for which to define top genes driving PCs}

\item{order}{ordering by which to plot the in heatmap of top genes driving PCs}
}
\description{
Plot the results obtained from PCA analysis as cell embedding in 2D space and annotation of gene loadings
}
