% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read.10x.data}
\alias{read.10x.data}
\title{Read 10X data}
\usage{
read.10x.data(path)
}
\arguments{
\item{path}{Path to directory containing matrix.mtx, genes.tsv, and barcodes.tsv}
}
\value{
A matrix of genes by cells
}
\description{
Load sparse data matrices from 10X genomics.
}
