% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read.preclustered.datasets}
\alias{read.preclustered.datasets}
\title{Read Preclustered Datasets}
\usage{
read.preclustered.datasets(environment, path = NA, recursive = T,
  rerun = F)
}
\arguments{
\item{environment}{\code{environment} object}

\item{path}{search path for previous projects}

\item{recursive}{recursive path search}

\item{rerun}{whether to rerun the reading process or load from cache}
}
\description{
Read previous analysis of multiple datasets to perform integrated analysis.
}
\examples{
\donttest{
pooled_env <- initialize.project(datasets = c("LCMV1", "LCMV2"),
origins = rep("CD44+ cells", 2),
experiments = c("Rep1", "Rep2"),
data.path = "~/LCMV/",
work.path = "~/LCMV/LCMV_analysis")
pooled_env <- read.preclustered.datasets(pooled_env)
}
}
