% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data2.R
\name{read_10x_data}
\alias{read_10x_data}
\title{Read 10X Data}
\usage{
read_10x_data(path, subsample = NULL, seed = 1)
}
\arguments{
\item{path}{Path to directory containing matrix.mtx, genes.tsv, and barcodes.tsv}

\item{subsample}{number of cells to subsample}

\item{seed}{seed for subsampling of cells}
}
\value{
a SingleCellExperiment object
}
\description{
Load sparse data matrices from 10X genomics into a SingleCellExperiment object.
}
