% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ribosomal.score}
\alias{ribosomal.score}
\title{Compute ribosomal score}
\usage{
ribosomal.score(environment, control = T, knn = 10)
}
\arguments{
\item{environment}{The environment object}

\item{control}{TODO}

\item{knn}{Number of nearest neighbor}
}
\value{
A vector of ribosonal score (TODO)
}
\description{
TODO
}
