% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess2.R
\name{ribosomal_score}
\alias{ribosomal_score}
\title{Compute Ribosomal Score}
\usage{
ribosomal_score(sce, control = T, knn = 10, verbose = F, force = F)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{control}{whether to subtract the score defined by technically similar genes}

\item{knn}{number of nearest neighbor}

\item{verbose}{Whether to print diagnostic messages}

\item{force}{Whether to rerun and overwrite already computed ribosomal score}
}
\value{
a vector of ribosomal genes activation score
}
\description{
Compute the activation level of ribosomal genes.
}
