% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE.R
\name{run.tSNE}
\alias{run.tSNE}
\title{Run tSNE}
\usage{
run.tSNE(environment, perplexity, max_iter, rerun)
}
\arguments{
\item{environment}{The environment}

\item{perplexity}{The perplexity parameter of tSNE}

\item{max_iter}{The maximum number of iterations to run the tSNE}

\item{rerun}{Whether to rerun}
}
\description{
TODO
}
