% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{set.cluster.names}
\alias{set.cluster.names}
\title{Set cluster names}
\usage{
set.cluster.names(environment, names)
}
\arguments{
\item{environment}{The environment object}

\item{names}{The cluster names}
}
\value{
The environment object with cluster named from provided input
}
\description{
TODO
}
