% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_LCMV1_example}
\alias{setup_LCMV1_example}
\title{Set up LCMV1 example}
\usage{
setup_LCMV1_example()
}
\value{
An environment object containing the LCMV1_small data
}
\description{
Set the path in the LCMV1_small object
}
