% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_LCMV_example}
\alias{setup_LCMV_example}
\title{Set up LCMV1 example}
\usage{
setup_LCMV_example(dataset = "LCMV1")
}
\arguments{
\item{dataset}{The name of dataset (LCMV1 or LCMV2)}
}
\value{
An environment object containing the LCMV1_small data
}
\description{
Set the path in the LCMV object
}
\examples{
LCMV1 <- setup_LCMV_example("LCMV1")
}
