% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_pooled_env}
\alias{setup_pooled_env}
\title{Set up the pooled environment}
\usage{
setup_pooled_env()
}
\value{
An environment object containing the two LCMV datasets
}
\description{
Set the path in the LCMV object
}
\examples{
pooled_env <- setup_pooled_env()
}
