% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA2.R
\name{shuffled_PCA}
\alias{shuffled_PCA}
\title{Parallelized PCA Analysis}
\usage{
shuffled_PCA(sce, regress = NULL, groups = NULL, nShuffleRuns = 10,
  threshold = 0.1, maxPCs = 100, label = NULL, mem = "2GB",
  time = "0:10:00", rerun = F, local = F)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object}

\item{regress}{gene signature activation scores to regress}

\item{groups}{experimental design annotation to guide dataset-specific regression}

\item{nShuffleRuns}{number of shuffled analyses}

\item{threshold}{FDR threshold}

\item{maxPCs}{maximum number of possible PCs}

\item{label}{optional analyses label folder}

\item{mem}{HPC memory}

\item{time}{HPC time}

\item{rerun}{whether to rerun the analysis rather than load from cache}

\item{local}{whether to run jobs locally on slurm instead of submitting the job}

\item{clear.previously.calculated.clustering}{whether to clear previous clustering analysis}
}
\value{
\code{environment} parameter containing PC coordinates
}
\description{
Run PCA analysis with a simulation analysis of shuffled data to determine the appropriate number of PCs.
}
