% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{summarize}
\alias{summarize}
\title{Summarize the clustering results}
\usage{
summarize(environment, perplexity = seq(10, 30, 10), max_iter = 10000,
  rerun = F, robust = F, order = NA, contrast = "all",
  min.fold = 1.5, quantile = 0.95)
}
\arguments{
\item{environment}{The environment object}

\item{perplexity}{The perplexity parameter for tSNE anlaysis}

\item{max_iter}{Maximum iterations for tSNE}

\item{rerun}{Whether to rerun}

\item{robust}{TODO}

\item{order}{TODO}

\item{contrast}{TODO}

\item{min.fold}{TODO}

\item{quantile}{TODO}
}
\value{
TODO
}
\description{
TODO
}
