% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data2.R
\name{update_summary_stats}
\alias{update_summary_stats}
\title{Compute summary statistics on count matrix}
\usage{
update_summary_stats(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
A data frame containing the number of UMI, genes with expression,
mean ribosomal gene counts, mean mitochondrial gene counts, fraction of
ribosomal gene expression per cell, fraction of mitochrondrial gene expression per cell
}
\description{
Compute summary statistics on count matrix
}
