% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{visualize.cluster.cors.heatmaps}
\alias{visualize.cluster.cors.heatmaps}
\title{Visualize cluster correlation heatmap}
\usage{
visualize.cluster.cors.heatmaps(environment, work.path, similarity)
}
\arguments{
\item{environment}{The environment object}

\item{work.path}{TODO}

\item{similarity}{TODO}
}
\value{
TODO
}
\description{
TODO
}
