% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{visualize.cluster.similarity.stats}
\alias{visualize.cluster.similarity.stats}
\title{Plot Similarity Results}
\usage{
visualize.cluster.similarity.stats(environment, similarity)
}
\arguments{
\item{environment}{\code{environment} object}

\item{similarity}{similarity matrix defined in compare.cluster.similarity or get.robust.cluster.similarity}
}
\description{
Perform hierarchical clustering and plot cluster similarities according to dendrogram.
}
\examples{
\donttest{
cluster.similarity <- assess.cluster.similarity(pooled_env)
similarity <- cluster.similarity$similarity
map <- cluster.similarity$map
filtered.similarity <- get.robust.cluster.similarity(
    pooled_env, similarity, min.sd = qnorm(.9), max.q.val = 0.01, rerun = F)
visualize.cluster.cors.heatmaps(pooled_env, pooled_env$work.path, filtered.similarity)
visualize.cluster.similarity.stats(pooled_env, similarity)
}
}
