% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{add.confounder.variables}
\alias{add.confounder.variables}
\title{Add Confounder Variables}
\usage{
add.confounder.variables(environment, ...)
}
\arguments{
\item{environment}{\code{environment} object}

\item{...}{vectors containing activation levels of confounding variables}
}
\value{
\code{environment} parameter containing added confounder variable
}
\description{
Add confounder variables' activation level per cell to \code{environment} object.
}
\examples{
\donttest{
LCMV1 <- setup_LCMV_example()
LCMV1 <- add.confounder.variables(LCMV1, ribosomal.score = ribosomal.score(LCMV1))
}
}
