% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simiarlity.R
\name{assess.cluster.similarity}
\alias{assess.cluster.similarity}
\title{Assess Cluster Similarity}
\usage{
assess.cluster.similarity(environment,
  diff.exp.file = "main.datasets.diff.exp.rds",
  cluster.similarity.function = pearson.correlation, label = "pearson",
  rerun = F)
}
\arguments{
\item{environment}{\code{environment} object}

\item{diff.exp.file}{name of differential expression results file}

\item{cluster.similarity.function}{which similarity function to use (either 'pearson.correlation' or '?') Mamie - there was another similarity function using euclidean distance. Do you know where did it go to? Can you replace the '?' with the name of this other function?}

\item{label}{name of the similarity measure to use for the results folder}

\item{rerun}{whether to rerun the analysis or load from cache}
}
\value{
pairwise cluster similarity measures
}
\description{
Assess similarity between pairs of clusters.
}
\examples{
\donttest{
data.path <- system.file("extdata", package = "robustSingleCell")
pooled_env <- initialize.project(datasets = c("LCMV1", "LCMV2"),
origins = rep("CD44+ cells", 2),
experiments = c("Rep1", "Rep2"),
data.path = data.path)
pooled_env <- read.preclustered.datasets(pooled_env)
pooled_env <- add.confounder.variables(
    pooled_env,
    ribosomal.score = ribosomal.score(pooled_env),
    mitochondrial.score = mitochondrial.score(pooled_env),
    cell.cycle.score = cell.cycle.score(pooled_env),
    Exhaustion = controlled.mean.score(pooled_env, exhaustion_markers))
pooled_env <- PCA(pooled_env, clear.previously.calculated.clustering = F)
summarize(pooled_env, contrast = "datasets")
cluster.similarity <- assess.cluster.similarity(pooled_env)
}
}
