% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{cluster.analysis}
\alias{cluster.analysis}
\title{Distributed Clustering Analysis}
\usage{
cluster.analysis(environment, knn.ratios = c(0.01, 0.05, 0.1),
  nShuffleRuns = 10, shuffledKNN = 10, loadPreviousKnn = T,
  rerun = F, deleteCache = F, mem = "4GB", time = "0:15:00",
  plot = T)
}
\arguments{
\item{environment}{\code{environment} object}

\item{knn.ratios}{range of KNN parameters to scan (corresponding to different resolutions)}

\item{nShuffleRuns}{number of shuffled clustering analyses to perform per KNN threshold}

\item{loadPreviousKnn}{whether to load previous analysis results}

\item{rerun}{whether to rerun the analysis rather than load from cache}

\item{mem}{HPC memory}

\item{time}{HPC time}

\item{plot}{whether to plot the clustering qualities compared to shuffled}

\item{shuffleKNN}{number of closest KNN shuffled analyses to include in background clustering quality computation}

\item{deleteCacheWhether}{whether to delete cache files}
}
\value{
\code{environment} parameter containing clustering assignment and provisional cluster names
}
\description{
Perform clustering analysis for a range of hyperparameter (KNN Ratios) values and assess clustering quality relative to simulation analysis of shuffled data.
}
