% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simiarlity.R
\name{compare.cluster.similarity}
\alias{compare.cluster.similarity}
\title{Compare the similarity between clusters}
\usage{
compare.cluster.similarity(environment,
  diff.exp.file = "main.datasets.diff.exp.RData",
  cluster.similarity.function = pearson.correlation, label = "pearson",
  rerun = F)
}
\arguments{
\item{environment}{The environment object}

\item{diff.exp.file}{TODO}

\item{cluster.similarity.function}{TODO}

\item{label}{TOOD}

\item{rerun}{TODO}
}
\value{
TODO
}
\description{
TODO
}
