% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{controlled.mean.score}
\alias{controlled.mean.score}
\title{Compute controled mean score}
\usage{
controlled.mean.score(environment, genes, knn = 10,
  exclude.missing.genes = T, constrain.cell.universe = NA)
}
\arguments{
\item{environment}{THe environment object}

\item{genes}{TODO}

\item{knn}{Number of nearest neighbors}

\item{exclude.missing.genes}{Whether to exclude genes with missing values}

\item{contrain.cell.universe}{TODO}
}
\value{
TODO
}
\description{
TODO
}
