% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get.cluster.names}
\alias{get.cluster.names}
\title{Get names of the clusters}
\usage{
get.cluster.names(environment, types, min.fold = 1.25, max.Qval = 0.1,
  print = T)
}
\arguments{
\item{environment}{The environment object}

\item{types}{TODO}

\item{min.fold}{TODO}

\item{max.Qval}{TODO}

\item{print}{TODO}
}
\value{
TODO
}
\description{
TODO
}
