% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simiarlity.R
\name{get.robust.cluster.similarity}
\alias{get.robust.cluster.similarity}
\title{Get Robust Cluster Similarity}
\usage{
get.robust.cluster.similarity(environment, similarity,
  min.sd = qnorm(0.95), max.q.val = 0.01, rerun = F)
}
\arguments{
\item{environment}{\code{environment} object}

\item{similarity}{pearson correlation between clusters' FC vectors defined in compare.cluster.similarity}

\item{min.sd}{minimum standard deviation for cluster reproducibility assessment}

\item{max.q.val}{maximum q value for cluster correlation cutoff}

\item{rerun}{whether to rerun the analysis or load from cache}
}
\value{
filtered cluster similarity matrix
}
\description{
Use cross-replicate experiment cluster similarity to remove irreproducible clusters.
}
