% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get.variable.genes}
\alias{get.variable.genes}
\title{Get highly variable genes by Heteroscedasticity controlled binning}
\usage{
get.variable.genes(environment, min.mean = 0.05, min.frac.cells = 0,
  min.dispersion.scaled = 1, rerun = F)
}
\arguments{
\item{min.mean}{Minimum mean expression per gene}

\item{min.frac.cells}{Minimum fraction of cells expressing each gene}

\item{min.dispersion.scaled}{Minimum dispersion value}

\item{rerun}{Whether to rerun}
}
\value{
\code{environment} parameter containing highly variable genes selection
}
\description{
Get highly variable genes by Heteroscedasticity controlled binning
}
