% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{mitochondrial.score}
\alias{mitochondrial.score}
\title{Compute Mitochondrial Score}
\usage{
mitochondrial.score(environment, control = F, knn = 10)
}
\arguments{
\item{environment}{\code{environment} object}

\item{control}{whether to subtract the score defined by technically similar genes}

\item{knn}{number of nearest neighbor}
}
\value{
a vector of mitochondrial genes activation score
}
\description{
Compute the activation level of mitochondrial genes.
}
\examples{
LCMV1 <- setup_LCMV_example()
mitochondrial.score <- mitochondrial.score(LCMV1)
}
