% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Parallelized PCA}
\usage{
PCA(environment, regress = NA, groups = NA, nShuffleRuns = 10,
  threshold = 0.1, maxPCs = 100, label = NA, mem = "2GB",
  time = "0:10:00", rerun = F,
  clear.previously.calculated.clustering = T)
}
\arguments{
\item{environment}{The environment object}

\item{regress}{Gene signature activation scores to regress}

\item{groups}{Experimental design annotation to guide dataset-specific regression}

\item{nShuffleRuns}{Number of shuffled analyses}

\item{threshold}{FDR threshold}

\item{maxPCs}{Maximum number of possible PCs}

\item{label}{Optional analyses label folder}

\item{mem}{HPC memory}

\item{time}{HPC time}

\item{rerun}{Whether to rerun}

\item{clear.previously.calculated.clustering}{Whether to clear previous clustering analysis}
}
\value{
TODO
}
\description{
TODO
}
