% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.PCA}
\alias{plot.PCA}
\title{Plot PCA}
\usage{
\method{plot}{PCA}(environment, quantile, order)
}
\arguments{
\item{environment}{The environment object}

\item{quantile}{quantile}

\item{order}{order}
}
\description{
TODO: description
}
