% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read.data}
\alias{read.data}
\title{Read data from 10X files or a raw data matrix and perform normalization, QC filtering and duplicates removal}
\usage{
read.data(environment, genome = "mm10", min.genes.per.cell = 500,
  max.genes.per.cell.quantile = 0.98,
  max.UMIs.per.cell.quantile = 0.98, min.cells.per.gene = 1,
  max.mitochondrial.frac = 0.1, max.ribosomal.frac = NA,
  cell.filters = NA, raw.data.matrices = NA, rerun = F,
  subsample = NULL, seed = 0)
}
\arguments{
\item{genome}{Genome annotation}

\item{min.genes.per.cell}{Minimum required number of genes per cell}

\item{max.genes.per.cell.quantile}{Upper quantile for number of genes per cell}

\item{max.UMIs.per.cell.quantile}{Upper quantile for number of UMIs per cell}

\item{min.cells.per.gene}{Minimum required number of cells per gene}

\item{max.mitochondrial.frac}{Maximum fraction of reads mapped to mitochondrial
genes per cell}

\item{max.ribosomal.frac}{Maximum fraction of reads mapped to ribosomal genes per cell}

\item{cell.filters}{Filtering option for cells based on marker genes}

\item{raw.data.matrices}{Logical indicating if data matrices is provided instead of 10X dataset}

\item{rerun}{Whether to rerun loading the dataset}

\item{subsample}{Number of cells to subsample}

\item{seed}{seed for subsampling}
}
\description{
Read data from 10X files or a raw data matrix and perform normalization, QC filtering and duplicates removal
}
