% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read.preclustered.datasets}
\alias{read.preclustered.datasets}
\title{Read preclustered datasets}
\usage{
read.preclustered.datasets(environment, path = NA, recursive = T,
  rerun = F)
}
\arguments{
\item{path}{Search path for previous projects}

\item{recursive}{Recursive path search}

\item{rerun}{Whether to rerun the reading process}
}
\description{
Read preclustered datasets
}
