% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_10x_data}
\alias{read_10x_data}
\title{Read 10X Data}
\usage{
read_10x_data(path)
}
\arguments{
\item{path}{Path to directory containing matrix.mtx, genes.tsv, and barcodes.tsv}
}
\value{
a matrix of genes by cells
}
\description{
Load sparse data matrices from 10X genomics.
}
