% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ribosomal.score}
\alias{ribosomal.score}
\title{Compute Ribosomal Score}
\usage{
ribosomal.score(environment, control = T, knn = 10)
}
\arguments{
\item{environment}{\code{environment} object}

\item{control}{whether to subtract the score defined by technically similar genes}

\item{knn}{number of nearest neighbor}
}
\value{
a vector of ribosomal genes activation score
}
\description{
Compute the activation level of ribosomal genes.
}
\examples{
LCMV1 <- setup_LCMV_example()
ribosomal.score <- ribosomal.score(LCMV1)
}
