% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE.R
\name{run.tSNE}
\alias{run.tSNE}
\title{Perform tSNE Analyses}
\usage{
run.tSNE(environment, perplexity, max_iter, rerun)
}
\arguments{
\item{environment}{\code{environment} object}

\item{perplexity}{perplexity parameter of tSNE}

\item{max_iter}{maximum number of iterations to run the tSNE}

\item{rerun}{whether to rerun or load from cache}
}
\value{
Distributed job identified object
}
\description{
Run distributed tSNE analysis for multiple input hyperparameters
}
