% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE.R
\name{run_tSNE}
\alias{run_tSNE}
\title{Perform tSNE Analyses}
\usage{
run_tSNE(environment, perplexity, max_iter, rerun, local = F,
  mem = "4GB", time = "0:15:00")
}
\arguments{
\item{environment}{\code{environment} object}

\item{perplexity}{perplexity parameter of tSNE}

\item{max_iter}{maximum number of iterations to run the tSNE}

\item{rerun}{whether to rerun or load from cache}

\item{local}{whether to run tSNE locally}

\item{mem}{Memory for each job; default 4 GB}

\item{time}{Time for each job; default 15 minutes}
}
\value{
Distributed job identified object
}
\description{
Run distributed tSNE analysis for multiple input hyperparameters
}
