% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{set.cluster.names}
\alias{set.cluster.names}
\title{Set Cluster Names in Environment}
\usage{
set.cluster.names(environment, names)
}
\arguments{
\item{environment}{\code{environment} object}

\item{names}{cluster names defined in get.cluster.names}
}
\value{
\code{environment} object coded with cluster names
}
\description{
Save the cluster names in storage and in the \code{environment} object
}
