% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{summarize}
\alias{summarize}
\title{Summarize the clustering results into figures}
\usage{
summarize(environment, perplexity = seq(10, 30, 10), max_iter = 10000,
  rerun = F, order = NA, contrast = "all", min.fold = 1.5,
  quantile = 0.95)
}
\arguments{
\item{environment}{The environment object}

\item{perplexity}{The perplexity parameter for tSNE anlaysis}

\item{max_iter}{Maximum iterations for tSNE}

\item{rerun}{Whether to rerun}

\item{order}{Order in which to plot the clusters}

\item{contrast}{Either 'all' indicating differential expression between one cluster against all others or 'datasets' indicating differential expression analysis comparing one cluster to all other within each dataset separately ('datasets' should be used in pooled analysis for optimal results)}

\item{min.fold}{Minimum fold change for filtering final differentially expressed gene lists}

\item{quantile}{q-value cutoff}
}
\description{
Summarize the clustering results into figures
}
