% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{visualize.cluster.cors.heatmaps}
\alias{visualize.cluster.cors.heatmaps}
\title{Visualize Correlation}
\usage{
visualize.cluster.cors.heatmaps(environment, work.path, similarity)
}
\arguments{
\item{environment}{\code{environment} object}

\item{work.path}{where to locate the figures}

\item{similarity}{similarity matrix defined in compare.cluster.similarity or get.robust.cluster.similarity}
}
\description{
Plot correlation heatmaps for each pair of datasets.
}
\examples{
\dontrun{
cluster.similarity <- assess.cluster.similarity(pooled_env)
similarity <- cluster.similarity$similarity
map <- cluster.similarity$map
filtered.similarity <- get.robust.cluster.similarity(
    pooled_env, similarity, min.sd = qnorm(.9), max.q.val = 0.01, rerun = F)
visualize.cluster.cors.heatmaps(pooled_env, pooled_env$work.path, filtered.similarity)
}
}
