% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{visualize.cluster.similarity.stats}
\alias{visualize.cluster.similarity.stats}
\title{Plot Similarity Results}
\usage{
visualize.cluster.similarity.stats(environment, similarity)
}
\arguments{
\item{environment}{\code{environment} object}

\item{similarity}{similarity matrix defined in compare.cluster.similarity or get.robust.cluster.similarity}
}
\description{
Perform hierarchical clustering and plot cluster similarities according to dendrogram.
}
